+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import Content from '@/app/components/404/Content';
+import Content from '@/components/Page/NotFound/Content';
 
 export default function NotFound() {
   return <Content />;
