+++ src/components/Page/Archive/Pagination/index.ts
@@ -1,0 +1,2 @@
+export { Pagination } from './Pagination';
+export { usePagination, DOTS } from './usePagination';
