+++ src/components/Page/Home/Hero/index.tsx
