+++ src/components/Page/Post/Mokuji/type.ts
